# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastShip API V1",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "operationId": "7df33efb604b635d7b5ebd4812b854bf",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all news"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/detail": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get detail news by ID",
                "operationId": "c27c6cc20d215288f91b547c9e122654",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of the news",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News details"
                    },
                    "404": {
                        "description": "News not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notification": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get all notification",
                "operationId": "cdf9a4ffea19ee90a2ec814b97df760b",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type(system, news, promotion, order,transaction)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all notifications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notification/detail": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Get detail notification by ID",
                "operationId": "9382e2b99082942cdc53db0a1234ff1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification details"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notification/delete": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Delete notification",
                "operationId": "460ba3fe45c5e7d5610b02dca30a033d",
                "requestBody": {
                    "description": "Delete notification",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID notification",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delete successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notification/read_all": {
            "post": {
                "tags": [
                    "Notification"
                ],
                "summary": "Read all notification",
                "operationId": "22bbef85300d36fc84c44ad1e4b54d79",
                "requestBody": {
                    "description": "Read all notification",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Read all successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}